% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{bar}
\alias{bar}
\title{Create a Bar Chart from a Data Frame}
\usage{
bar(data, category_col, value_col)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{category_col}{The name of the column to use for the categories (x-axis).}

\item{value_col}{The name of the column to use for the values (y-axis).}
}
\value{
A ggplot object representing the bar chart.
}
\description{
Creates a bar chart from a given data frame.
}
