% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_inventory_data}
\alias{cell_inventory_data}
\title{Example Cell Culture Inventory Data}
\format{
A tibble with 7 rows and 6 variables:
\describe{
  \item{CellLine}{The name of the cell line (character)}
  \item{PassageNumber}{The passage number of the cells in the vial (numeric)}
  \item{VialCount}{The number of vials available for that batch (numeric)}
  \item{Viability}{The post-thaw cell viability percentage (numeric)}
  \item{FreezeDate}{The date the vials were frozen (Date)}
  \item{Researcher}{The name of the researcher who froze the cells (character)}
}
}
\source{
Fictional data created for the MACCI package.
}
\usage{
cell_inventory_data
}
\description{
A toy dataset containing a fictional inventory of cryopreserved cell vials,
useful for demonstrating the functionality of the MACCI package.
}
\keyword{datasets}
