% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{save_to_excel}
\alias{save_to_excel}
\title{Save a Data Frame to an Excel File}
\usage{
save_to_excel(data, file_path)
}
\arguments{
\item{data}{The data frame to save.}

\item{file_path}{The path and name for the new Excel file (e.g., "sorted_data.xlsx").}
}
\value{
Invisibly returns the path to the new file.
}
\description{
Writes a data frame to a new .xlsx file.
}
