% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scatter}
\alias{scatter}
\title{Create a Scatter Plot}
\usage{
scatter(data, x_col, y_col)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{x_col}{The name of the column for the x-axis.}

\item{y_col}{The name of the column for the y-axis.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
Generates a scatter plot from a data frame using ggplot2.
}
\examples{
df <- data.frame(
  PassageNumber = c(5, 8, 12, 15),
  Viability = c(99.0, 95.5, 92.1, 89.8)
)
scatter(df, x_col = "PassageNumber", y_col = "Viability")
}
