% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{decreasing}
\alias{decreasing}
\title{Sort a Data Frame in Decreasing Order}
\usage{
decreasing(data, sort_by_col)
}
\arguments{
\item{data}{A data frame to sort.}

\item{sort_by_col}{The name of the column to sort by.}
}
\value{
A data frame sorted in decreasing order.
}
\description{
Sorts a data frame in descending order based on a specified column.
}
