% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization.R
\name{increasing}
\alias{increasing}
\title{Sort a Data Frame in Increasing Order}
\usage{
increasing(data, sort_by_col)
}
\arguments{
\item{data}{A data frame to sort.}

\item{sort_by_col}{The name of the column to sort by.}
}
\value{
A data frame sorted in increasing order.
}
\description{
Sorts a data frame in ascending order based on a specified column.
}
