% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{pie}
\alias{pie}
\title{Create a Pie Chart from a Data Frame}
\usage{
pie(data, category_col, value_col)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{category_col}{The name of the column to use for the slices.}

\item{value_col}{The name of the column to use for the slice sizes.}
}
\value{
A ggplot object representing the pie chart.
}
\description{
Creates a pie chart from a given data frame.
}
