% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{scatter}
\alias{scatter}
\title{Create a Scatter Plot from a Data Frame}
\usage{
scatter(data, x_col, y_col)
}
\arguments{
\item{data}{A data frame containing the data to plot.}

\item{x_col}{The name of the column for the x-axis.}

\item{y_col}{The name of the column for the y-axis.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
Creates a scatter plot from a given data frame.
}
